requirements.txt
fastapi
uvicorn[standard]
chromadb
ollama
python-docx
pypdf
python-multipart
streamlit
requests
python-dotenv
